review updates

+++ kie-wb-common-screens/kie-wb-common-server-ui/kie-wb-common-server-ui-client/src/main/java/org/kie/workbench/common/screens/server/management/client/container/status/ContainerRemoteStatusPresenter.java
@@ -1,17 +1,18 @@
 /*
- * Copyright 2015 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2016 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
-*/
+ */
 
 package org.kie.workbench.common.screens.server.management.client.container.status;
 
@@ -19,20 +20,19 @@ import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 import javax.annotation.PostConstruct;
+import javax.annotation.PreDestroy;
 import javax.enterprise.context.Dependent;
 import javax.enterprise.event.Observes;
 import javax.inject.Inject;
 
 import com.google.gwt.user.client.ui.IsWidget;
-import org.jboss.errai.ioc.client.container.IOC;
 import org.kie.server.api.model.KieContainerStatus;
 import org.kie.server.controller.api.model.events.ServerInstanceDeleted;
 import org.kie.server.controller.api.model.events.ServerInstanceUpdated;
 import org.kie.server.controller.api.model.runtime.Container;
 import org.kie.server.controller.api.model.spec.ContainerSpec;
 import org.kie.workbench.common.screens.server.management.client.container.status.card.ContainerCardPresenter;
-
-import static org.uberfire.commons.validation.PortablePreconditions.*;
+import org.kie.workbench.common.screens.server.management.client.util.IOCUtil;
 
 @Dependent
 public class ContainerRemoteStatusPresenter {
@@ -45,14 +45,17 @@ public class ContainerRemoteStatusPresenter {
     }
 
     private final View view;
+    private final IOCUtil iocUtil;
 
     private final Map<String, Map<String, ContainerCardPresenter>> index = new HashMap<String, Map<String, ContainerCardPresenter>>();
 
     private ContainerSpec containerSpec;
 
     @Inject
-    public ContainerRemoteStatusPresenter( final View view ) {
+    public ContainerRemoteStatusPresenter( final View view,
+                                           final IOCUtil iocUtil ) {
         this.view = view;
+        this.iocUtil = iocUtil;
     }
 
     @PostConstruct
@@ -64,44 +67,48 @@ public class ContainerRemoteStatusPresenter {
     }
 
     public void onServerInstanceUpdated( @Observes final ServerInstanceUpdated serverInstanceUpdated ) {
-        checkNotNull( "serverInstanceUpdated", serverInstanceUpdated );
-        final String updatedServerInstanceKey = serverInstanceUpdated.getServerInstance().getServerInstanceId();
-        if ( index.containsKey( updatedServerInstanceKey ) ) {
-            final Map<String, ContainerCardPresenter> oldIndex = new HashMap<String, ContainerCardPresenter>( index.remove( updatedServerInstanceKey ) );
-            final Map<String, ContainerCardPresenter> newIndexIndex = new HashMap<String, ContainerCardPresenter>( serverInstanceUpdated.getServerInstance().getContainers().size() );
-            index.put( updatedServerInstanceKey, newIndexIndex );
-            for ( final Container container : serverInstanceUpdated.getServerInstance().getContainers() ) {
-                ContainerCardPresenter presenter = oldIndex.remove( container.getContainerName() );
-                if ( !container.getStatus().equals( KieContainerStatus.STOPPED ) ) {
-                    if ( presenter != null ) {
-                        presenter.updateContent( serverInstanceUpdated.getServerInstance(), container );
-                    } else {
-                        presenter = buildContainer( container );
+        if ( serverInstanceUpdated != null &&
+                serverInstanceUpdated.getServerInstance() != null ) {
+            final String updatedServerInstanceKey = serverInstanceUpdated.getServerInstance().getServerInstanceId();
+            if ( index.containsKey( updatedServerInstanceKey ) ) {
+                final Map<String, ContainerCardPresenter> oldIndex = new HashMap<String, ContainerCardPresenter>( index.remove( updatedServerInstanceKey ) );
+                final Map<String, ContainerCardPresenter> newIndexIndex = new HashMap<String, ContainerCardPresenter>( serverInstanceUpdated.getServerInstance().getContainers().size() );
+                index.put( updatedServerInstanceKey, newIndexIndex );
+                for ( final Container container : serverInstanceUpdated.getServerInstance().getContainers() ) {
+                    ContainerCardPresenter presenter = oldIndex.remove( container.getContainerName() );
+                    if ( !container.getStatus().equals( KieContainerStatus.STOPPED ) ) {
+                        if ( presenter != null ) {
+                            presenter.updateContent( serverInstanceUpdated.getServerInstance(), container );
+                        } else {
+                            presenter = buildContainer( container );
+                        }
+                        newIndexIndex.put( container.getContainerName(), presenter );
                     }
-                    newIndexIndex.put( container.getContainerName(), presenter );
                 }
-            }
-            for ( final ContainerCardPresenter presenter : oldIndex.values() ) {
-                presenter.delete();
-            }
-        } else {
-            for ( final Container container : serverInstanceUpdated.getServerInstance().getContainers() ) {
-                if ( container.getServerTemplateId().equals( containerSpec.getServerTemplateKey().getId() ) &&
-                        container.getContainerSpecId().equals( containerSpec.getId() ) ) {
-                    buildAndIndexContainer( container );
+                for ( final ContainerCardPresenter presenter : oldIndex.values() ) {
+                    presenter.delete();
+                }
+            } else {
+                for ( final Container container : serverInstanceUpdated.getServerInstance().getContainers() ) {
+                    if ( container.getServerTemplateId().equals( containerSpec.getServerTemplateKey().getId() ) &&
+                            container.getContainerSpecId().equals( containerSpec.getId() ) ) {
+                        buildAndIndexContainer( container );
+                    }
                 }
             }
         }
     }
 
     public void onDelete( @Observes final ServerInstanceDeleted serverInstanceDeleted ) {
-        checkNotNull( "serverInstanceDeleted", serverInstanceDeleted );
-        final String deletedServerInstanceId = serverInstanceDeleted.getServerInstanceId();
-        if ( index.containsKey( deletedServerInstanceId ) ) {
-            final Map<String, ContainerCardPresenter> oldIndex = index.remove( deletedServerInstanceId );
-            if ( oldIndex != null ) {
-                for ( final ContainerCardPresenter presenter : oldIndex.values() ) {
-                    presenter.delete();
+        if ( serverInstanceDeleted != null &&
+                serverInstanceDeleted.getServerInstanceId() != null ) {
+            final String deletedServerInstanceId = serverInstanceDeleted.getServerInstanceId();
+            if ( index.containsKey( deletedServerInstanceId ) ) {
+                final Map<String, ContainerCardPresenter> oldIndex = index.remove( deletedServerInstanceId );
+                if ( oldIndex != null ) {
+                    for ( final ContainerCardPresenter presenter : oldIndex.values() ) {
+                        presenter.delete();
+                    }
                 }
             }
         }
@@ -123,7 +130,7 @@ public class ContainerRemoteStatusPresenter {
     }
 
     private ContainerCardPresenter buildContainer( final Container container ) {
-        final ContainerCardPresenter cardPresenter = newCard();
+        final ContainerCardPresenter cardPresenter = iocUtil.newInstance( this, ContainerCardPresenter.class );
         cardPresenter.setup( container.getServerInstanceKey(), container );
         view.addCard( cardPresenter.getView().asWidget() );
         return cardPresenter;
@@ -137,8 +144,9 @@ public class ContainerRemoteStatusPresenter {
         index.get( container.getServerInstanceKey().getServerInstanceId() ).put( container.getContainerName(), cardPresenter );
     }
 
-    ContainerCardPresenter newCard() {
-        return IOC.getBeanManager().lookupBean( ContainerCardPresenter.class ).getInstance();
+    @PreDestroy
+    public void destroy() {
+        iocUtil.cleanup( this );
     }
 
 }
